Move sync command in it's own module.
